Update: Tree Style Tab Compatibility.
If a user disables the AppMenu button then restarts the menu tool-bar
disappears, Lets trigger it to be created, However the user may have to
manually toggle the menu-tool-bar to remove the draw delay from adding
it here.

Side Note: There seems to be a few large errors in Tree Style Tab with
the new search see browser console.