Adjusted: Custom Button Color.
Here we have added percentage control to the custom button color, Now
you can change the position of the 2 colors on the button.

We also added a 3 color option with percentage control for users who
wish to have finer grade of control in there custom button color, This
also allows smoother gradients.