Use elementHandler for targeting multiple element states.
Here we have created a function that uses the element ids as an array
with the element function type and state.

Usage example this.elementHandler(['id'], "type", state);
[id] is an in-line array
['ctraddon_pw_tabwidthinfo','ctraddon_pw_tabwidthinfo2','ctraddon_pw_tabwidthinfo3']
[type] is a string that accepts the following
(disabled|hidden|visibility)
[state] is ether a boolean or string, for an array of id's you want to
hide using the type 'hidden' must be boolean where as for an array of
id's with "visibility" must be of type string.
this is expandable (currently only using if 3 or more).