Reduce CTRpreferences.json file size by almost half.
Here we have reduced the export of CTRpreferences.json by almost half.

Here we also allow the import of the older larger generated
CTRpreferences.json.

In both cases only the import of CTR based preferences can be done, If a
non CTR preference manages to be imported it will be prefixed with the
CTR prefix making it invalid.